+++ qml.qrc
@@ -1,9 +1,5 @@
 <RCC>
     <qresource prefix="/">
-        <file>qml/Main.qml</file>
-        <file>qml/MenuItem.qml</file>
-        <file>extra-files/draft/icons/fingerterm.png</file>
-        <file>extra-files/draft/icons/power.png</file>
-        <file>extra-files/draft/icons/xochitl.png</file>
+        <file>Main.qml</file>
     </qresource>
 </RCC>
